requirements.txt
streamlit
Pillow
tensorflow
opencv-python
numpy
streamlit-option-menu
scikit-learn==1.2.2
